Crate jsonrpsee

jsonrpsee wrapper crate.

Optional features

The jsonrpsee crate composes JSON-RPC functionality behind optional feature flags to provide for client and server communication over specific protocols. There are no default features, all functionality must be opted in to accordingly. The following features are available.

http-client - JSON-RPC client functionality over HTTP protocol.

wasm-client - JSON-RPC client functionality over web-sys.

ws-client - JSON-RPC client functionality over WebSocket protocol.

macros - JSON-RPC API generation convenience by derive macros.

client-core - Enables minimal client features to generate the RPC API without transports.

client - Enables all client features including transports.

server-core - Enables minimal server features to generate the RPC API without transports.

server - Enables all server features including transports.

full - Enables all features.

async-client - Enables the async client without any transport.

client-ws-transport - Enables ws transport with TLS.

client-ws-transport-no-tls - Enables ws transport without TLS.

client-web-transport - Enables websys transport.

Re-exports

pub use jsonrpsee_server as server;

pub use tokio;

pub use jsonrpsee_proc_macros as proc_macros;

pub use tracing;

pub use jsonrpsee_types as types;

pub use jsonrpsee_core as core;

Modules

helpers

Helpers.

Macros

rpc_params

Convert the given values to a crate::params::ArrayParams as expected by a jsonrpsee Client (http or websocket).

Structs

BatchResponse

Serialized batch response.

BatchResponseBuilder

Builder to build a BatchResponse.

BoundedSubscriptions

This wraps tokio::sync::Semaphore and is used to limit the number of subscriptions per connection.

BoundedWriter

Bounded writer that allows writing at most max_len bytes.

ConnectionId

Connection ID.

DisconnectError

Error that may occur during crate::server::MethodSink::send or crate::server::SubscriptionSink::send.

Extensions

A type map of protocol extensions.

IsUnsubscribed

Represents a subscription until it is unsubscribed.

MethodResponse

Represents a response to a method call.

MethodResponseFuture

Future that resolves when the method response has been processed.

MethodResponseNotifyTx

Sends a message once the method response has been processed.

MethodSink

Sink that is used to send back the result to the server for a specific method.

Methods

Reference-counted, clone-on-write collection of synchronous and asynchronous methods.

PendingSubscriptionAcceptError

The error returned while accepting a subscription.

PendingSubscriptionSink

Represents a single subscription that is waiting to be accepted or rejected.

ResponsePayload

Similar to jsonrpsee_types::ResponsePayload but possible to with an async-like API to detect when a method response has been sent.

RpcModule

Sets of JSON-RPC methods can be organized into “module“s that are in turn registered on the server or, alternatively, merged with other modules to construct a cohesive API. RpcModule wraps an additional context argument that can be used to access data during call execution.

Subscription

Wrapper struct that maintains a subscription “mainly” for testing.

SubscriptionKey

Represent a unique subscription entry based on SubscriptionId and ConnectionId.

SubscriptionMessage

Subscription message.

SubscriptionSink

Represents a single subscription that hasn’t been processed yet.

SubscriptionState

Helper struct to manage subscriptions.

Enums

CallOrSubscription

This represent a response to a RPC call and Subscribe calls are handled differently because we want to prevent subscriptions to start before the actual subscription call has been answered.

MethodCallback

Callback wrapper that can be either sync or async.

MethodKind

The kind of the JSON-RPC method call, it can be a subscription, method call or unknown.

MethodResponseError

Method response error.

MethodResult

Result of a method, either direct value or a future of one.

MethodsError

The error that can occur when Methods::call or Methods::subscribe is invoked.

NotifyMsg

A message that that tells whether notification was succesful or not.

SendTimeoutError

Error that may occur during crate::server::MethodSink::send_timeout or crate::server::SubscriptionSink::send_timeout.

SubscriptionCloseResponse

Represents what action that will sent when a subscription callback returns.

SubscriptionMessageInner

A complete subscription message or partial subscription message.

TrySendError

Error that may occur during crate::server::MethodSink::try_send or crate::server::SubscriptionSink::try_send.

Traits

IntoResponse

Something that can be converted into a JSON-RPC method call response.

IntoSubscriptionCloseResponse

Convert something into a subscription close notification before a subscription is terminated.

Functions

batch_response_error

Create a JSON-RPC error response.

prepare_error

Figure out if this is a sufficiently complete request that we can extract an Id out of, or just plain unparseable garbage.

Type Aliases

AsyncMethod

Similar to SyncMethod, but represents an asynchronous handler.

MaxResponseSize

Max response size.

RawRpcResponse

Raw response from an RPC A tuple containing:

Subscribers

Type-alias for subscribers.

SubscriptionMethod

Method callback for subscriptions.

SubscriptionPermit

Subscription permit.

SyncMethod

A MethodCallback is an RPC endpoint, callable with a standard JSON-RPC request, implemented as a function pointer to a Fn function taking four arguments: the id, params, a channel the function uses to communicate the result (or error) back to jsonrpsee, and the connection ID (useful for the websocket transport).