Struct jsonrpsee::MethodSink

pub struct MethodSink { /* private fields */ }

Sink that is used to send back the result to the server for a specific method.

Implementations

impl MethodSink

pub fn new(tx: Sender<String>) -> MethodSink

Create a new MethodSink with unlimited response size.

pub fn new_with_limit(tx: Sender<String>, max_response_size: u32) -> MethodSink

Create a new MethodSink with a limited response size.

pub fn is_closed(&self) -> bool

Returns whether this channel is closed without needing a context.

pub async fn closed(&self)

Same as tokio::sync::mpsc::Sender::closed.

Cancel safety

This method is cancel safe. Once the channel is closed, it stays closed forever and all future calls to closed will return immediately.

pub const fn max_response_size(&self) -> u32

Get the max response size.

pub fn try_send(&mut self, msg: String) -> Result<(), TrySendError>

Attempts to send out the message immediately and fails if the underlying connection has been closed or if the message buffer is full.

Returns the message if the send fails such that either can be thrown away or re-sent later.

pub async fn send(&self, msg: String) -> Result<(), DisconnectError>

Async send which will wait until there is space in channel buffer or that the subscription is disconnected.

pub async fn send_error<'a>( &self, id: Id<'a>, err: ErrorObject<'a> ) -> Result<(), DisconnectError>

Send a JSON-RPC error to the client

pub async fn send_timeout( &self, msg: String, timeout: Duration ) -> Result<(), SendTimeoutError>

Similar to to MethodSink::send but only waits for a limited time.

pub async fn has_capacity(&self) -> Result<(), DisconnectError>

Waits for there to be space on the return channel.