Crate jsonrpsee

Jsonrpsee wrapper crate.

Optional features

The jsonrpsee crate composes JSON-RPC functionality behind optional feature flags to provide for client and server communication over specific protocols. There are no default features, all functionality must be opted in to accordingly. The following features are avaliable.

http-client - JSON-RPC client functionality over HTTP protocol.

http-server - JSON-RPC server functionality over HTTP protocol.

ws-client - JSON-RPC client functionality over WebSocket protocol.

ws-server - JSON-RPC server functionality over WebSocket protocol.

macros - JSON-RPC API generation convenience by derive macros.

client - Enables http-client and ws-client features.

server - Enables http-server and ws-server features.

full - Enables client, server and macros features.

Re-exports

pub use jsonrpsee_ws_client as ws_client;

pub use jsonrpsee_proc_macros as proc_macros;

pub use jsonrpsee_types as types;

Macros

rpc_params

JSON-RPC client convenience macro to build params.