Crate jsonrpsee

jsonrpsee wrapper crate.

Optional features

The jsonrpsee crate composes JSON-RPC functionality behind optional feature flags to provide for client and server communication over specific protocols. There are no default features, all functionality must be opted in to accordingly. The following features are available.

http-client - JSON-RPC client functionality over HTTP protocol.

wasm-client - JSON-RPC client functionality over web-sys.

ws-client - JSON-RPC client functionality over WebSocket protocol.

macros - JSON-RPC API generation convenience by derive macros.

client-core - Enables minimal client features to generate the RPC API without transports.

client - Enables all client features including transports.

server-core - Enables minimal server features to generate the RPC API without transports.

server - Enables all server features including transports.

full - Enables all features.

async-client - Enables the async client without any transport.

client-ws-transport - Enables ws transport with TLS.

client-ws-transport-no-tls - Enables ws transport without TLS.

client-web-transport - Enables websys transport.

Re-exports

pub use jsonrpsee_http_client as http_client;

pub use jsonrpsee_ws_client as ws_client;

pub use jsonrpsee_server as server;

pub use jsonrpsee_proc_macros as proc_macros;

pub use tracing;

pub use jsonrpsee_types as types;

pub use jsonrpsee_core as core;

Macros

rpc_params

Convert the given values to a crate::params::ArrayParams as expected by a jsonrpsee Client (http or websocket).

Structs

RpcModule

Sets of JSON-RPC methods can be organized into a “module“s that are in turn registered on the server or, alternatively, merged with other modules to construct a cohesive API. RpcModule wraps an additional context argument that can be used to access data during call execution.

SubscriptionSink

Represents a single subscription.