Crate jsonrpsee

jsonrpsee wrapper crate.

Optional features

The jsonrpsee crate composes JSON-RPC functionality behind optional feature flags to provide for client and server communication over specific protocols. There are no default features, all functionality must be opted in to accordingly. The following features are avaliable.

http-client - JSON-RPC client functionality over HTTP protocol.

http-server - JSON-RPC server functionality over HTTP protocol.

wasm-client - JSON-RPC client functionality over web-sys.

ws-client - JSON-RPC client functionality over WebSocket protocol.

ws-server - JSON-RPC server functionality over WebSocket protocol.

macros - JSON-RPC API generation convenience by derive macros.

client - Enables http-client and ws-client features.

server - Enables http-server and ws-server features.

full - Enables client, server and macros features.

async-client - Enables the async client without any transport.

client-ws-transport - Enables ws transport with TLS.

client-ws-transport-no-tls - Enables ws transport without TLS.

Re-exports

pub use jsonrpsee_http_server as http_server;

pub use jsonrpsee_ws_server as ws_server;

pub use jsonrpsee_proc_macros as proc_macros;

pub use tracing;

pub use jsonrpsee_types as types;

pub use jsonrpsee_core as core;

Structs

PendingSubscription

Represent a pending subscription which waits until it’s either accepted or rejected.

RpcModule

Sets of JSON-RPC methods can be organized into a “module“s that are in turn registered on the server or, alternatively, merged with other modules to construct a cohesive API. RpcModule wraps an additional context argument that can be used to access data during call execution.

SubscriptionSink

Represents a single subscription.